refactored considering evgeniia uml

+++ src/main/java/com/gorshkov/shop/model/Product.java
@@ -3,9 +3,9 @@ package com.gorshkov.shop.model;
 public class Product {
     private final int id;
     private final String name;
-    private final int price;
+    private final double price;
 
-    public Product(int id, String name, int price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -19,7 +19,7 @@ public class Product {
         return name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
